// src/app/models/user.ts
export class IUser {
    CNIC: string;
    address: string;
    banner: string;
    bio: string;
    birthday: string;
    city: string;
    country: string;
    email: string;
    favourite_places: string;
    first_name: string;
    gender: string;
    id: string;
    is_authenticated: string;
    is_online: string;
    last_name: string;
    latitude: string;
    location_name: string;
    longitude: string;
    passion: string;
    phone: string;
    picture: string;
    user_id: string;
    username: string;
    wishlist: string;
}